Some comments
This is kind of an annoying commit as I have not contributed any codes but I have some important questions and explanations and suggestions.
I think the most important two are the todo.md file and the in-code usage of TODO tags. We should write general goals in the markdown file (as that can, and should, include much more detail) and we can put a flag on the code comment so that we don't forget what to add and what to remove in the long run.
I am not creating a new branch for this commit for this is purely for comments and to determine our way of collaboration but that should be our general practice. Otherwise we loose old versions of our code and everything becomes a mess...