understanding cluster
I'm trying to write a couple of functions to be able to hop between
clusters but I have yet to properly understand it... This commit
includes just a printing method for the routing table.
We are going to use that table to navigate in the cluster and between
clusters we should use a path finding algorithm.