rating calculations are almost done
wow, it has been a long day!
Added parameters:
    wQ, wR, wT are for if we want to use the weighted average of
quality, rarity and timeliness in the rating calculation
    potency and consistency (see below)
AttackerType
    I think this is important. It is better and easier to follow than
just using strings to denote the attacker type of nodes. A PROBLEM (or
imperfection) arises about benevolent nodes: I have included a
BENEVOLENT case for this enum just as a default, but I don't think this
enum type should ever be called if we do not explicitly know that the
node is malicious.
    calculateMalRating is for when a malicious node decides to vote "as
it wishes", so not according to our expectations (which is calculated
with the normal calculateRating method)

Potency & Consistency
    These two are attributes of an IoTNode. BE VERY CAREFUL NOT TO
REASSIGN THEM. I have tried to use const everytime I reached them with a
method but I am too dumb to know how to declare them as const with
OMNeT.

Quality
    We "normally" calculate this with potency and consistency. Since the
quality of the service is written in the message, it makes sense that
this is the "actual" quality of the message (instead of a possible
"perceived quality" or sth). So it makes perfect sense to define a
"calculateQuality" method which calls the respective function to
generate a subpar quality. (I guess this is not too clear but whatever)

calculateRating & sendRating
    The latter was updated to have the rating to be sent as a second
parameter with the first parameter being the providerID, which makes
sense.
    The former should take quality, timeliness and rarity as its
parameters. It should not have to rely on the nodes themselves!!!

Timeliness
    I dunno how to do this yet....