a function for modularization and minor git stuff
propagateRatingReplica is defined to make handleServiceRatingMsg a bit
shorter.
I intend to continue on this trend.
We *should not* commit results files, they just bloat our commits with
out any benefit; but I am afraid to remove anything from the repo at
this point.
We *must not* commit BlockchainMEssage_m.h for it is the root of all evil.