Attack Type from init, Pot and Cons
Set the attacker type from omnet.init by changing the related variable
to the assigned int value. The list is written as a comment, I could not
have it written as a string. (Actually I could but it would be too
cumbersome)

Also Potency and Consistency values are now set according to a uniform
distribution, and they should not be changed elsewhere, as we have
written calcuateRating and calcQuality functions in a way that handles
bad behaviour and there is no need to change these attributes for
malicious nodes.