Ability-TrustScore graph
It looks like I did much more than I actually did,
because of my linter.
It reformats some lines, sorry.

Added the "static" keyword to some variables and functions
to make the ODR error shut up.

Included a script in the results file that generates the current graph.
Perhaps we will want to rewrite it, so that there is a greater unison
amongst our figures in the paper.